# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Origin Agent Platform Product API
  version: v1.0-alpha
  description: >
    Product-level API contract for Origin Agent Platform. The product owns this
    contract; server-side services implement it; PloyKit modules consume it.
...
servers:
  - url: https://api.origin.example
    description: Production API endpoint.
  - url: http://127.0.0.1:8080
    description: Local implementation endpoint.
  - url: http://127.0.0.1:4010
    description: Local contract/mock endpoint.
security:
  - serviceToken: []
    originSignature: []
    originTimestamp: []
    originClaims: []
...
paths:
...
  /v1/agents:
    get:
      tags: [Agents]
      operationId: listAgents
      summary: List agents.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Agent list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
    post:
      tags: [Agents]
      operationId: createAgent
      summary: Create an agent.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreateRequest"
      responses:
        "201":
          description: Agent created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /v1/agents/{agent_id}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
    get:
      tags: [Agents]
      operationId: getAgent
      summary: Get an agent.
      responses:
        "200":
          description: Agent resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Agents]
      operationId: patchAgent
      summary: Patch an agent.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentPatchRequest"
      responses:
        "200":
          description: Agent patched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Agents]
      operationId: deleteAgent
      summary: Delete an agent.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: Agent deleted.
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/agents/{agent_id}/versions:
    get:
      tags: [Agents]
      operationId: listAgentVersions
      summary: List agent release and draft versions.
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Agent version list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentVersionListResponse"
...
  /v1/service-connections/{connection_id}/resources:
    get:
      tags: [ServiceConnections]
      operationId: listServiceConnectionResources
      summary: List resources discovered from a service connection.
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "200":
          description: Discovered connection resources.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceConnectionResourceListResponse"
...
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: service-token
    originSignature:
      type: apiKey
      in: header
      name: x-origin-agentops-signature
    originTimestamp:
      type: apiKey
      in: header
      name: x-origin-agentops-timestamp
    originClaims:
      type: apiKey
      in: header
      name: x-origin-agentops-claims
  parameters:
...
    IdempotencyKey:
      name: idempotency-key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
...
    ConnectionId:
      name: connection_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Conflict, usually idempotency or state transition related.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Meta:
      type: object
      required: [request_id, trace_id]
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        workspace_id:
          type: string
        tenant_id:
          type: string
      additionalProperties: true
    ListMeta:
      allOf:
        - $ref: "#/components/schemas/Meta"
        - type: object
          properties:
            next_cursor:
              type: string
              nullable: true
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          $ref: "#/components/schemas/Meta"
...
    Agent:
      type: object
      required: [agent_id, name, status, created_at, updated_at]
      properties:
        agent_id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, validating, review_required, published, disabled, archived]
        version:
          type: string
        identifier:
          type: string
          description: Human-friendly stable slug used in product UI.
        workspace_id:
          type: string
        version_policy:
          type: string
        prompt:
          type: string
        prompt_profile:
          type: object
          additionalProperties: true
        skill_ids:
          type: array
          items:
            type: string
        tool_ids:
          type: array
          items:
            type: string
        provider_ids:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        objectives:
          type: array
          items:
            type: string
        draft_version:
          type: string
          nullable: true
        last_run_at:
          type: string
          format: date-time
          nullable: true
        last_run_status:
          type: string
          nullable: true
        tool_count:
          type: integer
        run_count:
          type: integer
        eval_summary:
          type: object
          additionalProperties: true
        governance_summary:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        prompt:
          type: string
        identifier:
          type: string
        workspace_id:
          type: string
        version_policy:
          type: string
        skill_ids:
          type: array
          items:
            type: string
        tool_ids:
          type: array
          items:
            type: string
        provider_ids:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        objectives:
          type: array
          items:
            type: string
        approval_required:
          type: boolean
        external_api_allowed:
          type: boolean
        tracing_enabled:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
    AgentPatchRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        prompt:
          type: string
        identifier:
          type: string
        workspace_id:
          type: string
        version_policy:
          type: string
        status:
          type: string
        skill_ids:
          type: array
          items:
            type: string
        tool_ids:
          type: array
          items:
            type: string
        provider_ids:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        objectives:
          type: array
          items:
            type: string
        approval_required:
          type: boolean
        external_api_allowed:
          type: boolean
        tracing_enabled:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
...
    AgentResponse:
      type: object
      required: [agent, meta]
      properties:
        agent:
          $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/Meta"
    AgentListResponse:
      type: object
      required: [agents, meta]
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/ListMeta"
...
    AgentVersion:
      type: object
      required: [version, status, created_at]
      properties:
        version:
          type: string
        status:
          type: string
          enum: [draft, canary, stable, archived]
        release_notes:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        published_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
...
    AgentVersionListResponse:
      type: object
      required: [versions, meta]
      properties:
        versions:
          type: array
          items:
            $ref: "#/components/schemas/AgentVersion"
        meta:
          $ref: "#/components/schemas/ListMeta"
...
    ServiceConnectionResource:
      type: object
      required: [resource_id, name, method, path]
      properties:
        resource_id:
          type: string
        name:
          type: string
        method:
          type: string
        path:
          type: string
        purpose:
          type: string
        enabled:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
...
    ServiceConnectionResourceListResponse:
      type: object
      required: [resources, meta]
      properties:
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ServiceConnectionResource"
        meta:
          $ref: "#/components/schemas/ListMeta"
...